Change package name from `aggregates` to `aggregate`.
