+++ api-java/src/main/java/org/spine3/examples/todolist/c/aggregate/MismatchHelper.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.c.aggregates;
+package org.spine3.examples.todolist.c.aggregate;
 
 import com.google.protobuf.Timestamp;
 import org.spine3.change.StringMismatch;
