+++ api-java/src/main/java/org/spine3/examples/todolist/c/aggregate/TaskAggregateRoot.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.c.aggregates;
+package org.spine3.examples.todolist.c.aggregate;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.spine3.examples.todolist.TaskDefinition;
