+++ api-java/src/main/java/org/spine3/examples/todolist/c/aggregate/TaskLabelsPart.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.c.aggregates;
+package org.spine3.examples.todolist.c.aggregate;
 
 import com.google.protobuf.Message;
 import org.spine3.examples.todolist.LabelId;
@@ -40,10 +40,10 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static org.spine3.examples.todolist.c.aggregates.FailureHelper.throwCannotAssignLabelToTaskFailure;
-import static org.spine3.examples.todolist.c.aggregates.FailureHelper.throwCannotRemoveLabelFromTaskFailure;
-import static org.spine3.examples.todolist.c.aggregates.TaskFlowValidator.isValidAssignLabelToTaskCommand;
-import static org.spine3.examples.todolist.c.aggregates.TaskFlowValidator.isValidRemoveLabelFromTaskCommand;
+import static org.spine3.examples.todolist.c.aggregate.FailureHelper.throwCannotAssignLabelToTaskFailure;
+import static org.spine3.examples.todolist.c.aggregate.FailureHelper.throwCannotRemoveLabelFromTaskFailure;
+import static org.spine3.examples.todolist.c.aggregate.TaskFlowValidator.isValidAssignLabelToTaskCommand;
+import static org.spine3.examples.todolist.c.aggregate.TaskFlowValidator.isValidRemoveLabelFromTaskCommand;
 
 /**
  * The aggregate managing the state of a {@link TaskLabels}.
