+++ api-java/src/main/java/org/spine3/examples/todolist/context/EventBusEnricher.java
@@ -28,9 +28,9 @@ import org.spine3.examples.todolist.TaskDefinition;
 import org.spine3.examples.todolist.TaskDetails;
 import org.spine3.examples.todolist.TaskId;
 import org.spine3.examples.todolist.TaskLabel;
-import org.spine3.examples.todolist.c.aggregates.LabelAggregate;
-import org.spine3.examples.todolist.c.aggregates.TaskDefinitionPart;
-import org.spine3.examples.todolist.c.aggregates.TaskLabelsPart;
+import org.spine3.examples.todolist.c.aggregate.LabelAggregate;
+import org.spine3.examples.todolist.c.aggregate.TaskDefinitionPart;
+import org.spine3.examples.todolist.c.aggregate.TaskLabelsPart;
 import org.spine3.examples.todolist.repositories.LabelAggregateRepository;
 import org.spine3.examples.todolist.repositories.TaskDefinitionRepository;
 import org.spine3.examples.todolist.repositories.TaskLabelsRepository;
