+++ api-java/src/test/java/org/spine3/examples/todolist/c/aggregate/TaskLabelsPartTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.c.aggregates;
+package org.spine3.examples.todolist.c.aggregate;
 
 import com.google.common.base.Throwables;
 import com.google.protobuf.Message;
